Permit non-leading digits in NSID name segment
Fixes the failing tests with the [updated](https://github.com/bluesky-social/atproto-website/pull/402) example NSIDs from the spec.

This was a minimal change (just adding the `0-9`s), but i think the name-segment matcher could be further simplified:

```
(\.[a-zA-Z]([a-zA-Z0-9]{0,61}[a-zA-Z0-9])?)  # current
(\.[a-zA-Z][a-zA-Z0-9]{0,62})  # simpler, equivalent except for the internal capture group
```

but i will leave deciding that to maintainer taste.