trim https-well-known-method handle response (#322)
Handle resolution failed for me when it encountered http://benharri.org/.well-known/atproto-did , which currently has a trailing newline character.

It's not *supposed* to have a trailing newline, but the spec permits stripping it

> It is acceptable to strip any prefix and suffix whitespace from the response body before attempting to parse as a DID.

note that stripping a small amount may become a "SHOULD" in the spec, see https://bsky.app/profile/bnewbold.net/post/3lwezl4ztfs2x

either way i think this fix is at least still correct behaviour, but with more compatibility of DIDs encountered in the wild.